+++ llm_openai/tests/__init__.py
@@ -1,2 +1,3 @@
 from . import test_stream_metadata
 from . import test_reasoning_effort
+from . import test_effort_retry